requirements.txt
requests==2.32.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
scikit-learn==1.6.0